refactor(evals): promote viz ES|QL evaluators into @kbn/evals
Share validity/execution/bind-params/normalize/calibrated FuncEq from the
framework, drop redundant withEvaluatorSpan wrappers, and omit the standalone
validity evaluator from the viz suite default set (execution already AST-checks).
