+++ x-pack/platform/packages/shared/agent-builder-visualizations/kbn-evals-suite-agent-builder-visualizations/README.md
@@ -9,11 +9,12 @@ It drives the full `/api/agent_builder/converse` flow, extracts the ES|QL backin
 Per [issue #277136](https://github.com/elastic/kibana/issues/277136), "correct" for visualization ES|QL is not the same as for analytical ES|QL. This first increment covers:
 
 - **Visualization skill activated** (`CODE`) — the request loaded the visualization skill and called `platform.core.create_visualization` (guards against the agent answering with raw ES|QL / a table instead of a rendered visualization).
-- **ES|QL Validity** (`CODE`) — the generated query parses via `@kbn/esql-language`.
-- **ES|QL Execution Validity** (`CODE`) — the query executes against real sample data and (per example) returns rows. This is the tier that surfaces the fast-model regressions that motivated the suite.
-- **ES|QL Functional Equivalence** (`LLM` judge) — the generated query is equivalent to the ground-truth query. Uses a calibrated three-point rubric (`equivalent` / `equivalent_with_caveats` / `not_equivalent`) with explicit allow/deny lists, ported from `@kbn/evals-suite-security-esql-generation-regression`, so cosmetic differences (aliases, interchangeable bucketing, `?_tstart`/`?_tend` vs literal ranges) earn partial credit rather than a hard 0.
+- **ES|QL Execution Validity** (`CODE`, from `@kbn/evals`) — AST parse + execute against real sample data and (per example) return rows. This is the tier that surfaces the fast-model regressions that motivated the suite.
+- **ES|QL Functional Equivalence** (`LLM` calibrated judge, from `@kbn/evals`) — three-point rubric (`equivalent` / `equivalent_with_caveats` / `not_equivalent`) with cosmetic allow-lists for aliases, bucketing, and `?_tstart`/`?_tend` bounds.
 - **Trajectory** — the agent routed the request to `load_skill` → `platform.core.create_visualization`.
 - **Trace-based** — tokens / latency / tool-call counts from OTel spans.
+
+A standalone ES|QL Validity evaluator also lives in `@kbn/evals` (`createEsqlValidityEvaluator`) but is not in the default set here — execution already covers AST validation.
 
 Not yet covered (tracked as follow-up increments in the issue): chart-compatible-result (does the ES|QL result shape fit the requested chart type), chart-type-vs-intent, Lens/Vega-Lite config validity, and an MLLM visual-fidelity judge.
 
@@ -34,7 +35,7 @@ Seed examples live inline in `evals/visualization_creation/visualization_creatio
 - **Categorical / metric** golds include the raw-`@timestamp` time filter (`WHERE @timestamp >= ?_tstart AND @timestamp < ?_tend`).
 - **Time-series** golds express the window via the auto-bucket-count form (`BUCKET(@timestamp, 75, ?_tstart, ?_tend)` / `TBUCKET(75, ?_tstart, ?_tend)`); an extra `@timestamp` WHERE is optional and stripped before equivalence scoring.
 
-This keeps gold and candidate structurally parallel so the equivalence evaluators measure real differences instead of cosmetic ones. The `?_tstart` / `?_tend` bind params substitute to a **now-relative** window (see `src/evaluators/esql_bind_params.ts`), which brackets both `kibana_sample_data_logs` and the GCS snapshot replay data (whose timestamps are shifted to end at `now` by the replay pipeline).
+This keeps gold and candidate structurally parallel so the equivalence evaluators measure real differences instead of cosmetic ones. The `?_tstart` / `?_tend` bind params substitute to a **now-relative** window (`getDefaultTimeBounds` / `substituteEsqlBindParams` in `@kbn/evals`), which brackets both `kibana_sample_data_logs` and the GCS snapshot replay data (whose timestamps are shifted to end at `now` by the replay pipeline).
 
 ### OTel data fixture
 
@@ -46,7 +47,3 @@ The snapshot contains OTel Demo logs, metrics, and traces. After first replay, c
 GET metrics-*/_field_caps?fields=system.cpu.*,@timestamp
 GET _data_stream/metrics-*
 ```
-
-## Notes
-
-The ES|QL validity, execution, and calibrated functional-equivalence evaluators plus the `?_tstart` / `?_tend` bind-param substitution are copied from `@kbn/evals-suite-security-esql-generation-regression` because Kibana module-visibility rules forbid importing across sibling `functional-tests` suites. If a third suite needs them, consider promoting them into `@kbn/evals` (which already exports the v1 `createEsqlEquivalenceEvaluator`).
