+++ x-pack/platform/packages/shared/agent-builder-visualizations/kbn-evals-suite-agent-builder-visualizations/moon.yml
@@ -20,12 +20,9 @@ dependsOn:
   - '@kbn/agent-builder-common'
   - '@kbn/core'
   - '@kbn/es-snapshot-loader'
-  - '@kbn/esql-language'
   - '@kbn/evals'
   - '@kbn/inference-common'
-  - '@kbn/inference-prompt-utils'
   - '@kbn/tooling-log'
-  - '@kbn/zod'
 tags:
   - functional-tests
   - package
