+++ x-pack/platform/packages/shared/kbn-evals/index.ts
@@ -167,7 +167,21 @@ export type {
   GroundTruthExtractor,
   RetrievedDoc,
 } from './src/evaluators/rag/types';
-export { createEsqlEquivalenceEvaluator } from './src/evaluators/esql';
+export {
+  createEsqlEquivalenceEvaluator,
+  createCalibratedEsqlEquivalenceEvaluator,
+  createEsqlExecutionEvaluator,
+  createEsqlValidityEvaluator,
+  getDefaultTimeBounds,
+  substituteEsqlBindParams,
+  normalizeEsqlForEquivalence,
+  stripRedundantTimestampBindBounds,
+  ESQL_EQUIVALENCE_EVALUATOR_NAME,
+  ESQL_CALIBRATED_EQUIVALENCE_EVALUATOR_NAME,
+  ESQL_CALIBRATED_EQUIVALENCE_JUDGE_VERSION,
+  ESQL_EXECUTION_EVALUATOR_NAME,
+  ESQL_VALIDITY_EVALUATOR_NAME,
+} from './src/evaluators/esql';
 
 export { createTrajectoryEvaluator } from './src/evaluators/trajectory';
 export { createConversationCoherenceEvaluator } from './src/evaluators/conversation_coherence';
