+++ x-pack/platform/packages/shared/kbn-evals/moon.yml
@@ -40,6 +40,7 @@ dependsOn:
   - '@kbn/es-errors'
   - '@kbn/es'
   - '@kbn/dot-text'
+  - '@kbn/esql-language'
 tags:
   - test-helper
   - package
