+++ x-pack/platform/packages/shared/kbn-evals/src/evaluators/esql/normalize_for_equivalence.ts
